+++ application/config/config.php
@@ -7,7 +7,7 @@ $config['menu_choices'] = array(
 
         array('name' => 'Home', 'link' => '/'),
         array('name' => 'Fleet', 'link' => '/fleet'),
-        array('name' => 'Flights', 'link' => '/charlie/brown'),
+        array('name' => 'Flights', 'link' => '/flights'),
         )
     );